Add mediatR
