+++ src/Eventus.Samples.Web/Controllers/BankAccountController.cs
@@ -1,34 +1,41 @@
 ﻿using System;
+using System.Threading.Tasks;
+using Eventus.Samples.Web.Features.BankAccount.DTO;
+using Eventus.Samples.Web.Features.BankAccount.Query;
+using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Eventus.Samples.Web.Controllers
 {
     public class BankAccountController : Controller
     {
+        private readonly IMediator _mediator;
+
+        public BankAccountController(IMediator mediator)
+        {
+            _mediator = mediator;
+        }
+
         /// <summary>
         /// Get bank account summary
         /// </summary>
         /// <param name="id">Bank Account ID</param>
         /// <response code="200">Returns the bank account summary</response>
-        [HttpGet("/bankaccount/{id:guid}")]
+        [HttpGet("api/bankaccount/{id:guid}")]
         [ProducesResponseType(typeof(BankAccountSummary), 200)]
-        public IActionResult GetAccount(Guid id)
+        public async Task<IActionResult> GetAccount(Guid id)
         {
-            return new ObjectResult(new BankAccountSummary
+            var account = await _mediator.Send(new GetBankAccountQuery
             {
-                Id = id,
-                AccountName = "Joe",
-                CurrentBalance = 100
-            });
+                Id = id
+            }).ConfigureAwait(false);
+
+            if (account == null)
+            {
+                return NotFound();
+            }
+
+            return new ObjectResult(account);
         }
-    }
-
-    public class BankAccountSummary
-    {
-        public Guid Id { get; set; }
-
-        public string AccountName { get; set; }
-
-        public int CurrentBalance { get; set; }
     }
 }
