+++ src/Eventus.Samples.Web/Eventus.Samples.Web.csproj
@@ -17,6 +17,8 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="FluentEmail.Mailgun" Version="2.1.0" />
+    <PackageReference Include="MediatR" Version="3.0.1" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="2.0.0" />
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.0.0" />
     <PackageReference Include="Microsoft.AspNetCore" Version="1.1.2" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="1.1.2" />
