+++ src/Eventus.Samples.Web/Features/BankAccount/DTO/BankAccountSummary.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Eventus.Samples.Web.Features.BankAccount.DTO
+{
+    public class BankAccountSummary
+    {
+        public Guid Id { get; set; }
+
+        public string AccountName { get; set; }
+
+        public int CurrentBalance { get; set; }
+    }
+}
