+++ src/Eventus.Samples.Web/Features/BankAccount/Query/GetBankAccountQuery.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Eventus.Samples.Web.Features.BankAccount.DTO;
+using MediatR;
+
+namespace Eventus.Samples.Web.Features.BankAccount.Query
+{
+    public class GetBankAccountQuery : IRequest<BankAccountSummary>
+    {
+        public Guid Id { get; set; }
+    }
+}
